+++ CHANGELOG.md
@@ -19,6 +19,7 @@
   * change: config: SDK is now installed in $HOME/.aspera/sdk instead of $HOME/.aspera/ascli/sdk
   * change(break): aoc/node: Simplification: gen4 operations: show modify permission thumbnail are now directly under node gen 4 command. Command `file` is suppressed. Option `path` is suppressed. The default expected argument is a path. To provide a file id, use selector syntax: %id:_file_id_
   * change(break): node: option `token_type` is removed, as starting with 4.3 basic token is only allowed with access keys, so use gen4 operations: `acc do self`
+  * fix: faspex5: Receive package from shared inbox
 
 * 4.13.0